backend/requirements.txt
django
langchain
langchain-community
pymupdf
psycopg2-binary
pillow
django-cors-headers
chromadb
langchain-chroma
sentence-transformers
langchain-huggingface
langchain-google-genai
python-dotenv
